refactor(llm): rename compaction modules to disclose their seam
compaction.ts -> session-summary.ts  (persists DB summary + prune)
conversation-compactor.ts -> context-budget.ts  (transient in-memory trim)

Update all import sites. No logic changes.
