+++ packages/server/src/llm/session-summary.ts
